// src/auth/enums/message-patterns.enum.ts
// message-patterns.enum.ts

export enum MessagePatternEnum {
  REGISTER = 'register',
  PHONE_LOGIN = 'phoneLogin',
  VERIFY_OTP = 'verifyOtp',
  LOGIN = 'login',
  RESET_PASSWORD = 'reset_password',
  FORGOT_PASSWORD = 'forgot_password',
  RESEND_EMAIL = 'resend_email',
  SOCIAL_LOGIN = 'social_login',
  VERIFY_ID = 'verify_id',
  PHONE_REGISTER = 'phone_register',
  SOCIAL_REGISTER = 'social_register',
  FIND_USER_BY_ID = 'find_user_by_id',
}